--- .idea/libraries/commons_lang_2_6.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="commons-lang-2.6">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/Downloads/commons-lang-2.6/commons-lang-2.6.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
