+++ build.gradle
@@ -1,0 +1,17 @@
+plugins {
+    id 'java'
+}
+
+group 'TCPClient'
+version '1.0-SNAPSHOT'
+
+sourceCompatibility = 1.8
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    compile group: 'org.apache.commons', name: 'commons-lang3', version: '3.9'
+    testCompile group: 'junit', name: 'junit', version: '4.12'
+}
